refactor: rm `RequireConfirmed` flag when finding output
since we know check for confirmation when choosing claimable swaps, we
dont have to do it a second time when creating the tx. this also causes
the old test case to stop working since we dont error and just ignore.
those can be safely removed since the confirmation check is now covered
elsewhere